[S12P11E206-253] [FE] Feat: GPS 좌표 기반 클라이밍 정보 불러오기 API 연동 구현

+++ front/lib/features/place/data/models/place_all_model.dart
@@ -1,7 +1,6 @@
-import 'package:big_decimal/big_decimal.dart';
 class PlaceAllModel {
-  final BigDecimal latitude;
-  final BigDecimal longitude;
+  final double latitude;
+  final double longitude;
 
   PlaceAllModel({
     required this.latitude,
